Zero + Mynt & Money On Chain integration

+++ contracts/masset/IBridge.sol
@@ -47,7 +47,7 @@ interface IBridge {
         address operator,
         address from,
         address to,
-        uint amount,
+        uint256 amount,
         bytes calldata userData,
         bytes calldata operatorData
     ) external;
